Improve OS independence of fuse_mount_sys().
Instead of littering it with #ifdefs, move the OS-dependent code to separate
functions (fuse_mount_retry() and fuse_mount_add()), and make the actual
mount() call be os_mount(), so that it can do something different on
platforms whose mount() arguments don't match Linux's.

For FreeBSD, simply remove the fstype argument in a #define.